switch to ALE for lsp
ALE comes with support for many language servers, and adding one or more
custom language servers is fairly easy and explanatory

at the time of making this commit, `rust-analyzer` is not available as
a built in ale language server. ale/pull/2837 is a pr that includes it.
a version of ALE with this pr included must be used for rust
autocomplete to work (additionally, rust-analyzer must be installed with
`ra_lsp_binary` reachable in $PATH